+++ lib/screens/transfer/contact/list.dart
@@ -1,0 +1,36 @@
+import 'package:flutter/material.dart';
+
+class ContactList extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text("Contacts"),
+      ),
+      body: ListView(
+        children: [
+          Card(
+            child: ListTile(
+              title: Text(
+                "Martha",
+                style: TextStyle(
+                  fontSize: 24.0,
+                ),
+              ),
+              subtitle: Text(
+                "1000",
+                style: TextStyle(
+                  fontSize: 16.0,
+                ),
+              ),
+            ),
+          )
+        ],
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {},
+        child: Icon(Icons.add),
+      ),
+    );
+  }
+}
