+++ lib/screens/transfer/dashboard.dart
@@ -1,6 +1,8 @@
 import 'package:bytebank/main.dart';
 import 'package:flutter/material.dart';
 
+import 'contact/list.dart';
+
 class Dashboard extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
